refactor(apps/web/인게임): useWaitingUIStore를 stores 폴더 아래로 이전
